% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detsgdp.R
\name{detsgdp}
\alias{detsgdp}
\title{detsgdp}
\usage{
detsgdp(gdp)
}
\arguments{
\item{gdp}{Time Series Object.}
}
\value{
graph showing the time series decomposition
}
\description{
Performs Ets decomposition and plots the results.
}
\examples{
detsgdp(gdp)

}
