% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffgdp.R
\name{diffgdp}
\alias{diffgdp}
\title{diffgdp}
\usage{
diffgdp(gdp)
}
\arguments{
\item{gdp}{Time Series Object.}
}
\value{
A differentiated Time Series Object
}
\description{
Takes the first difference of the time series object previously downloaded.
}
\examples{
diffgdp(gdp)

}
