% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstlgdp.R
\name{dstlgdp}
\alias{dstlgdp}
\title{dstlgdp}
\usage{
dstlgdp(gdp)
}
\arguments{
\item{gdp}{Time Series Object.}
}
\value{
graph showing the time series decomposition
}
\description{
Performs Stl decomposition and plots the results.
}
\examples{
dstlgdp(gdp)

}
