% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_arima.R
\name{forecast_arima}
\alias{forecast_arima}
\title{forecast_arima}
\usage{
forecast_arima(data = NULL)
}
\arguments{
\item{"data"}{The stock dataset that you want to use. If \code{NULL} selects TSLA  dataset.}
}
\value{
Graph showing forecasted values
}
\description{
Performs arima forecasting
}
\examples{
forecast_arima()
forecast_arima(aapl)

}
