% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastgdpets.R
\name{forecastgdpets}
\alias{forecastgdpets}
\title{forecastgdpets}
\usage{
forecastgdpets(gdp)
}
\arguments{
\item{gdp}{Time Series Object.}
}
\value{
graph showing the forecasted values
}
\description{
Performs forecast of the ets object and plots the results.
}
\examples{
forecastgdpets(gdp)

}
