% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastgdpstl.R
\name{forecastgdpstl}
\alias{forecastgdpstl}
\title{forecastgdpstl}
\usage{
forecastgdpstl(gdp)
}
\arguments{
\item{gdp}{Time Series Object.}
}
\value{
graph showing the forecasted values
}
\description{
Performs forecast of the stl and plots the results.
}
\examples{
forecastgdpstl(gdp)

}
