% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(id = "TSLA", from = "2012-09-01", to = Sys.Date())
}
\arguments{
\item{"id"}{character. Ticker of the desired stock. If \code{NULL} selects TSLA  dataset.}

\item{"from"}{date. The date you want to start. If \code{NULL} selects "2012-09-01"}

\item{"to"}{date. The date you want to stop. If \code{NULL} selects today's date}
}
\value{
A dataset of the selected stock
}
\description{
Retrieve stock dataset with date, open, close, high, low,volume and adjusted price.
}
\details{
You have to store the result in a variable in order to be able to perform further analisys
with the stock that you are intrested in
}
\examples{
get_data()
get_data("AAPL")
aapl <- get_data("AAPL")
stockname <- get_data("id")
}
