% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgdp.R
\name{getgdp}
\alias{getgdp}
\title{getgdp}
\usage{
getgdp()
}
\value{
A Time Series Object called gdp
}
\description{
Retrieve Fred GDP timeseries object anb stores it in a global variable called gdp.
}
\examples{
getgdp()

}
