% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laggdp.R
\name{laggdp}
\alias{laggdp}
\title{laggdp}
\usage{
laggdp(gdp)
}
\arguments{
\item{gdp}{Time Series Object.}
}
\value{
A lagged Time Series Object
}
\description{
Lags the time series object previously downloaded.
}
\examples{
laggdp(gdp)

}
