% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ema.R
\name{plot_ema}
\alias{plot_ema}
\title{plot_ema}
\usage{
plot_ema(data = NULL, end = Sys.Date())
}
\arguments{
\item{data}{dataframe. The stock dataset that you want to use. If \code{NULL} (the default) selects TSLA  dataset.}
}
\value{
A graph of the last 6 months price and EMA's
}
\description{
Plots the last 6 months with Exponential Moving Average
}
\examples{
plot_ema()
plot_ema(aapl)


}
