% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stockp.R
\name{plot_stockp}
\alias{plot_stockp}
\title{plot_data}
\usage{
plot_stockp(data = NULL, title = "")
}
\arguments{
\item{data}{dataframe. The stock dataset that you want to used. If \code{NULL} (the default) selects TSLA  dataset.}

\item{title}{character. title that you want to assign to the graph.}
}
\value{
A graph of the stock price
}
\description{
Plots a graph of the closing price vector.
}
\examples{
plot_stockp()
plot_stockp(aapl)
plot_stockp(tsla,TSLA)


}
