% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zoom.R
\name{plot_zoom}
\alias{plot_zoom}
\title{plot_zoom}
\usage{
plot_zoom(data = NULL, end = Sys.Date())
}
\arguments{
\item{"data"}{dataframe. The stock dataset that you want to use. If \code{NULL} (the default) selects TSLA dataset.}
}
\value{
A graph of the last 8 weeks
}
\description{
Plots last 8 weeks graph of the timeseries.
}
\examples{
plot_zoom()
plot_zoom(aapl)
}
