% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgdp.R
\name{plotgdp}
\alias{plotgdp}
\title{plotgdp}
\usage{
plotgdp(gdp)
}
\arguments{
\item{gdp}{Time Series Object.}
}
\value{
a graph of the time series
}
\description{
plots the time series.
}
\examples{
plot(gdp)

}
