% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_arima.R
\name{train_arima}
\alias{train_arima}
\title{train_arima}
\usage{
train_arima(data = NULL)
}
\arguments{
\item{data}{The stock dataset that you want to use. If \code{NULL} (the default) selects TSLA dataset.}
}
\value{
Graph showing the mean tendency of our forecasting over the test set close price move
}
\description{
Trains arima model
}
\examples{
train_arima()
train_arima(aapl)

}
