% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_NN.R
\name{forecast_NN}
\alias{forecast_NN}
\title{forecast_NN}
\usage{
forecast_NN(data = NULL)
}
\arguments{
\item{"data"}{The stock dataset that you want to use. If \code{NULL} (the default) selects TSLA  dataset.}
}
\value{
graph showing forecasted values
}
\description{
Performs a neural network forecasting
}
\examples{
forecast_NN()
forecast_NN(aapl)

}
